Split Auth0 code from rest of JWT
